split modules files
